requirements.txt
django
django_rest_framework
psycopg2-binary
djangorestframework-simplejwt
pillow
